[DIA-2061] new file for unit tests
